Create upload csv and append that to the sheet instead of the entire file
The scanner puts the data in stats as well as upload. So there is still
a hard copy of the data. When you click the upload button it takes the
data from upload and appends it to the google sheet and then deletes the
upload.csv. The header is only created if the stats csv is empty so we
do not get duplicates in the sheet.