Support dynamic spreadsheet id's
Before the upload and teams.csv were hardcoded, now they are pulled from
a properly formatted google sheet. The format is as follows

The sheet needs to be called "Config" in A1 is the teams.csv download
in A2 is your Google Spreadsheet ID that you want to upload too. In the
upload sheet it will upload to a Sheet called "RawData". NOTE: You can
have the config and rawdata be in the same spreadsheet you just need to
make the 2 sheets as described.

To get the Google Spreadsheet ID open the spreadsheet in your browser and
it is the string of letters/symbols/numbers after /d/

Example:
https://docs.google.com/spreadsheets/d/SPREADSHEET_ID/